@@ -6,7 +6,7 @@ import 'package:learn_ar_flutter/starting_page.dart';
 main() => runApp(MyApp());
 
 class MyApp extends StatelessWidget {
-  const MyApp({Key key}) : super(key: key);
+  const MyApp({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -17,15 +17,14 @@ class MyApp extends StatelessWidget {
     return MaterialApp(
       theme: ThemeData(
         primaryColor: Colors.deepPurple,
-        accentColor: Colors.red,
         textTheme: TextTheme(
-            bodyText1: TextStyle(color: Colors.deepPurple, fontSize: 50),
-            headline3: TextStyle(color: Colors.white, fontSize: 25)),
+            bodyLarge: TextStyle(color: Colors.deepPurple, fontSize: 50),
+            displaySmall: TextStyle(color: Colors.white, fontSize: 25)),
         dialogTheme: DialogTheme(
           backgroundColor: Colors.deepPurple,
           contentTextStyle: TextStyle(color: Colors.white, fontSize: 16),
         ),
-        scaffoldBackgroundColor: Colors.white,
+        scaffoldBackgroundColor: Colors.white, colorScheme: ColorScheme.fromSwatch().copyWith(secondary: Colors.red),
       ),
       home: StartingPage(),
       debugShowCheckedModeBanner: false,
